+++ hybracter/util.py
@@ -32,8 +32,19 @@ def print_version():
 def echo_click(msg, log=None):
     click.echo(msg, nl=False, err=True)
     if log:
-        with open(log, "a") as l:
-            l.write(msg)
+        # if log already exists
+        if os.path.exists(log):
+            with open(log, "a") as l:
+                l.write(msg)
+        # to create it
+        else:
+            directory_path = os.path.dirname(log)
+            # Create the directories recursively
+            os.makedirs(directory_path, exist_ok=True)
+            with open(log, "w") as l:
+                l.write(msg)
+
+
 
 
 def print_citation():
@@ -56,13 +67,6 @@ def msg_box(splash, errmsg=None, log=None):
     msg(("-" * (len(splash) + 4)), log=log)
     if errmsg:
         echo_click("\n" + errmsg + "\n", log=log)
-
-
-def default_to_ouput(ctx, param, value):
-    """Callback for --configfile; place value in output directory unless specified"""
-    if param.default == value:
-        return os.path.join(ctx.params["output"], value)
-    return value
 
 
 def default_to_ouput(ctx, param, value):
@@ -157,6 +161,7 @@ def run_snakemake(
 
     # if using a configfile
     if configfile:
+
         # copy sys default config if not present
         copy_config(configfile, log=log)